refactor gitea pullrequest
